added menu dropdown when start translating
